xAPI Reporter v 2-09 Windows 32 bit.zip is a 32 bit Windows stand alone
The standalone application is a LiveCode project that has been exported to a standalone runtime, you can download this file to Windows, double click to unzip and double click to run. It would be better if you can use the 64 bit version under Window 10, earlier versions of Windows may require this version to run.